feat: Introduce CustomIcon settings item (settings-ui)
This commit introduces a new `SettingsItem.CustomIcon` type to standardize the way custom icons are handled across different settings screens. This change consolidates the logic for selecting and displaying custom icons, improving code reusability and maintainability.

The `CustomIcon` composable has been extracted and now conforms to the `SettingsItem` interface. This allows it to be integrated seamlessly into the `SettingsItemContent` rendering system.

Affected files:
- ui/src/main/kotlin/com/eblan/launcher/ui/model/SettingsItem.kt
- ui/src/main/kotlin/com/eblan/launcher/ui/settings/Settings.kt
- feature/edit-application-info/src/main/kotlin/com/eblan/launcher/feature/editapplicationinfo/EditApplicationInfoScreen.kt
- feature/edit-grid-item/src/main/kotlin/com/eblan/launcher/feature/editgriditem/EditGridItemScreen.kt
