+++ ui/src/main/kotlin/com/eblan/launcher/ui/settings/Settings.kt
@@ -17,6 +17,10 @@
  */
 package com.eblan.launcher.ui.settings
 
+import android.content.Intent
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -24,21 +28,35 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import com.eblan.launcher.designsystem.component.VerticalSlideReveal
+import com.eblan.launcher.designsystem.icon.EblanLauncherIcons
+import com.eblan.launcher.domain.model.PackageManagerIconPackInfo
+import com.eblan.launcher.ui.R
 import com.eblan.launcher.ui.model.SettingsItem
+import com.eblan.launcher.common.R as commonR
 
 @Composable
 fun SettingsSwitch(
@@ -130,33 +148,6 @@ fun SettingsSwitch(
 @Composable
 fun SettingsColumn(
     modifier: Modifier = Modifier,
-    title: String,
-    subtitle: String,
-    onClick: () -> Unit,
-) {
-    Column(
-        modifier = modifier
-            .clickable(onClick = onClick)
-            .fillMaxWidth()
-            .padding(15.dp),
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleMedium,
-        )
-
-        Spacer(modifier = Modifier.height(5.dp))
-
-        Text(
-            text = subtitle,
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    }
-}
-
-@Composable
-fun SettingsColumn(
-    modifier: Modifier = Modifier,
     index: Int,
     size: Int,
     title: String,
@@ -193,6 +184,7 @@ fun SettingsColumn(
 
 @Composable
 fun SettingsItemContent(
+    modifier: Modifier = Modifier,
     settingsItem: SettingsItem,
     index: Int,
     size: Int,
@@ -200,6 +192,7 @@ fun SettingsItemContent(
     when (settingsItem) {
         is SettingsItem.Column -> {
             SettingsColumn(
+                modifier = modifier,
                 index = index,
                 size = size,
                 title = settingsItem.title,
@@ -210,6 +203,7 @@ fun SettingsItemContent(
 
         is SettingsItem.Switch -> {
             SettingsSwitch(
+                modifier = modifier,
                 index = index,
                 size = size,
                 checked = settingsItem.checked,
@@ -223,6 +217,7 @@ fun SettingsItemContent(
         is SettingsItem.CustomBackgroundColor,
         -> {
             CustomBackgroundColor(
+                modifier = modifier,
                 index = index,
                 size = size,
                 title = settingsItem.title,
@@ -233,12 +228,24 @@ fun SettingsItemContent(
 
         is SettingsItem.Row -> {
             SettingsRow(
+                modifier = modifier,
                 index = index,
                 size = size,
                 imageVector = settingsItem.imageVector,
                 title = settingsItem.title,
                 subtitle = settingsItem.subtitle,
                 onClick = settingsItem.onClick,
+            )
+        }
+
+        is SettingsItem.CustomIcon -> {
+            CustomIcon(
+                modifier = modifier,
+                customIcon = settingsItem.customIcon,
+                packageManagerIconPackInfos = settingsItem.packageManagerIconPackInfos,
+                onUpdateIconPackInfoPackageName = settingsItem.onUpdateIconPackInfoPackageName,
+                onUpdateUri = settingsItem.onUpdateUri,
+                onResetCustomIcon = settingsItem.onResetCustomIcon,
             )
         }
     }
@@ -254,6 +261,206 @@ fun SettingsCategoryText(
         text = text,
         style = MaterialTheme.typography.bodySmall,
     )
+}
+
+fun settingsItemShape(
+    index: Int,
+    size: Int,
+    radius: Dp = 16.dp,
+): Shape = when {
+    size == 1 -> RoundedCornerShape(size = radius)
+
+    index == 0 -> RoundedCornerShape(
+        topStart = radius,
+        topEnd = radius,
+        bottomStart = radius / 2,
+        bottomEnd = radius / 2,
+    )
+
+    index == size - 1 -> RoundedCornerShape(
+        topStart = radius / 2,
+        topEnd = radius / 2,
+        bottomStart = radius,
+        bottomEnd = radius,
+    )
+
+    else -> RoundedCornerShape(size = radius / 2)
+}
+
+@Composable
+private fun CustomIcon(
+    modifier: Modifier = Modifier,
+    customIcon: String?,
+    packageManagerIconPackInfos: List<PackageManagerIconPackInfo>,
+    onUpdateIconPackInfoPackageName: (
+        packageName: String,
+        label: String?,
+    ) -> Unit,
+    onUpdateUri: (String) -> Unit,
+    onResetCustomIcon: () -> Unit,
+) {
+    var expanded by remember { mutableStateOf(false) }
+
+    val context = LocalContext.current
+
+    val pickMedia =
+        rememberLauncherForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
+            if (uri != null) {
+                val flag = Intent.FLAG_GRANT_READ_URI_PERMISSION
+
+                context.contentResolver.takePersistableUriPermission(uri, flag)
+
+                onUpdateUri(uri.toString())
+            }
+        }
+
+    Column(modifier = modifier.fillMaxWidth()) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(15.dp),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            if (customIcon != null) {
+                AsyncImage(
+                    modifier = Modifier.size(40.dp),
+                    model = customIcon,
+                    contentDescription = null,
+                )
+            } else {
+                Icon(
+                    modifier = Modifier.size(40.dp),
+                    imageVector = EblanLauncherIcons.BrokenImage,
+                    contentDescription = null,
+                )
+            }
+
+            Spacer(modifier = Modifier.width(10.dp))
+
+            Column(modifier = Modifier.weight(1f)) {
+                Text(text = stringResource(R.string.custom_icon))
+
+                Spacer(modifier = Modifier.height(10.dp))
+
+                Text(text = customIcon ?: stringResource(commonR.string.none))
+            }
+
+            IconButton(
+                onClick = {
+                    expanded = !expanded
+                },
+            ) {
+                Icon(
+                    imageVector = if (expanded) {
+                        EblanLauncherIcons.ArrowDropUp
+                    } else {
+                        EblanLauncherIcons.ArrowDropDown
+                    },
+                    contentDescription = null,
+                )
+            }
+        }
+
+        VerticalSlideReveal(visible = expanded) {
+            Column(modifier = Modifier.fillMaxWidth()) {
+                SettingsColumn(
+                    title = stringResource(R.string.gallery),
+                    subtitle = stringResource(R.string.pick_icons_from_your_gallery),
+                    onClick = {
+                        pickMedia.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+                    },
+                )
+
+                packageManagerIconPackInfos.forEach { packageManagerIconPackInfo ->
+                    IconPackItem(
+                        icon = packageManagerIconPackInfo.icon,
+                        label = packageManagerIconPackInfo.label,
+                        packageName = packageManagerIconPackInfo.packageName,
+                        onClick = {
+                            onUpdateIconPackInfoPackageName(
+                                packageManagerIconPackInfo.packageName,
+                                packageManagerIconPackInfo.label,
+                            )
+                        },
+                    )
+                }
+
+                if (customIcon != null) {
+                    SettingsColumn(
+                        title = stringResource(R.string.reset_custom_icon),
+                        subtitle = stringResource(R.string.reset_custom_icon),
+                        onClick = onResetCustomIcon,
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun SettingsColumn(
+    modifier: Modifier = Modifier,
+    title: String,
+    subtitle: String,
+    onClick: () -> Unit,
+) {
+    Column(
+        modifier = modifier
+            .clickable(onClick = onClick)
+            .fillMaxWidth()
+            .padding(15.dp),
+    ) {
+        Text(
+            text = title,
+            style = MaterialTheme.typography.titleMedium,
+        )
+
+        Spacer(modifier = Modifier.height(5.dp))
+
+        Text(
+            text = subtitle,
+            style = MaterialTheme.typography.bodyMedium,
+        )
+    }
+}
+
+@Composable
+private fun IconPackItem(
+    modifier: Modifier = Modifier,
+    icon: ByteArray?,
+    label: String?,
+    packageName: String,
+    onClick: () -> Unit,
+) {
+    Row(
+        modifier = modifier
+            .clickable(onClick = onClick)
+            .fillMaxWidth()
+            .padding(15.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        AsyncImage(
+            modifier = Modifier.size(40.dp),
+            model = icon,
+            contentDescription = null,
+        )
+
+        Spacer(modifier = Modifier.width(20.dp))
+
+        Column {
+            Text(
+                text = label.toString(),
+                style = MaterialTheme.typography.titleMedium,
+            )
+
+            Spacer(modifier = Modifier.height(5.dp))
+
+            Text(
+                text = packageName,
+                style = MaterialTheme.typography.bodyMedium,
+            )
+        }
+    }
 }
 
 @Composable
@@ -302,28 +509,4 @@ private fun SettingsRow(
             }
         }
     }
-}
-
-fun settingsItemShape(
-    index: Int,
-    size: Int,
-    radius: Dp = 16.dp,
-): Shape = when {
-    size == 1 -> RoundedCornerShape(size = radius)
-
-    index == 0 -> RoundedCornerShape(
-        topStart = radius,
-        topEnd = radius,
-        bottomStart = radius / 2,
-        bottomEnd = radius / 2,
-    )
-
-    index == size - 1 -> RoundedCornerShape(
-        topStart = radius / 2,
-        topEnd = radius / 2,
-        bottomStart = radius,
-        bottomEnd = radius,
-    )
-
-    else -> RoundedCornerShape(size = radius / 2)
 }
